Fix windows build and add github CI (#26)
* Minor improvements to CMake code

- cmake_minimum_required must be before the project call
- Bump the minimum supported CMake version to something reasonable
  (CMake 3.13 is shipped in debian 10 buster)

* Fix cargo deprecation warnings

* Statically link LZ4 on windows

statically link into lz4 on windows to fix executables not linking on windows.
Presumably on Linux / darwin lz4 is available as a dynamic library.

a more proper solution might be to first search for the library
and then fallback to static linking only if
CMake can't find lz4.

* Reorder link-lib lines

brotli is linked into woff2 is linked into ots,
so the order of `-l` lines should be
1. ots
2. woff2
3. brotli

Some linkers may not care about the order and also scan right
to left (e.g. lld), but this is not a given.

* Add missing brotli files (missing symbols on windows)

add all enc files

* Add github CI workflow

* Update .github/workflows/test.yml

---------

Co-authored-by: Martin Robinson <mrobinson@igalia.com>