Replace vendored brotli with brotli-decompressor  (rust implementation) (#42)
* Use rust brotli implementation

brotli contains some large data tables,
and servo already uses the rust brotli crate,
so removing the vendored brotli allows us to reduce
the binary size of servo.

This requires us to create a new crate for woff2, since it
depends on brotli decompressor functions and we need to
specify this dependency, otherwise there will be linker errors.

* Mostly remove vendored brotli

Since we now use the brotli-decompressor crate,
we can remove the vendored brotli code.
We still need to keep the public header files,
since the woff2 build depends on those.
The public API should be stable though, as is
normally the cases for c libraries.

* Move brotli headers to fontsan-woff2

The brotli headers are only required for the woff2 build,
and need to be in the same directory as the package
for published builds to work.